Added Prepayment Invoice info blocks, changed version
