+++ etc/module.xml
@@ -1,5 +1,5 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../../lib/internal/Magento/Framework/Module/etc/module.xsd">
-    <module name="Heidelpay_Gateway" schema_version="17.7.19" setup_version="17.7.19">
+    <module name="Heidelpay_Gateway" schema_version="17.7.21" setup_version="17.7.21">
         <sequence>
             <module name="Magento_Sales" />
             <module name="Magento_Payment" />
