+++ view/adminhtml/templates/info/pdf/invoice.phtml
@@ -6,7 +6,7 @@
  * @author Stephano Vogel
  */
 /**
- * @var $block \Heidelpay\Gateway\Block\Info\InvoiceSecured
+ * @var $block \Heidelpay\Gateway\Block\Info\Invoice
  */
 ?>
 <?php echo $block->escapeHtml($block->getMethod()->getTitle()); ?>
