+++ view/adminhtml/templates/info/pdf/prepayment.phtml
@@ -6,7 +6,7 @@
  * @author Stephano Vogel
  */
 /**
- * @var $block \Heidelpay\Gateway\Block\Info\InvoiceSecured
+ * @var $block \Heidelpay\Gateway\Block\Info\Prepayment
  */
 ?>
 <?php echo $block->escapeHtml($block->getMethod()->getTitle()); ?>
