+++ view/frontend/templates/info/prepayment.phtml
@@ -1,0 +1,20 @@
+<?php
+/**
+ * @license Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
+ * @copyright Copyright © 2016-present Heidelberger Payment GmbH. All rights reserved.
+ * @link https://dev.heidelpay.de/magento
+ * @author Stephano Vogel
+ */
+/**
+ * @var $block \Heidelpay\Gateway\Block\Info\Prepayment
+ */
+?>
+<dl class="payment-method prepayment">
+    <dt class="title"><?php echo $block->escapeHtml($block->getMethod()->getTitle()) ?></dt>
+    <?php if ($block->getTransactionInfo() !== null) : ?>
+        <dd class="content">
+            <strong><?php echo $block->escapeHtml(__('Payment information')) ?></strong><br />
+            <?php echo $block->printAdditionalInformationHtml() ?>
+        </dd>
+    <?php endif; ?>
+</dl>
